chore: Rename utils -> bidi-mark-text, stringify -> compile-module
