+++ packages/messageformat/src/compile-module.js
